refactor: 선언부 분리
