+++ src/pages/AddCard/domain/domain.ts
@@ -1,4 +1,4 @@
-import { CardInfoInput, InputStatus } from '../../../type';
+import { AddCardFormProps, CardInfoInput, InputStatus } from '../../../type';
 import { INVALID_MESSAGE } from '../../../utils/constants';
 
 export const calcMultipleStatus = (arr: InputStatus[]): InputStatus => {
@@ -9,4 +9,40 @@ export const calcMultipleStatus = (arr: InputStatus[]): InputStatus => {
 
 export const getErrorMessage = (inputType: CardInfoInput, status: InputStatus) => {
   return INVALID_MESSAGE[inputType][status];
+};
+
+export const getTotalStatus = (formInputs: AddCardFormProps) => {
+  return Object.values(formInputs).map((input) => input.status);
+};
+
+export const getSubmitData = ({
+  cardCompany,
+  cardFirstNumber,
+  cardSecondNumber,
+  cardThirdNumber,
+  cardFourthNumber,
+  cardOwner,
+  expireMonth,
+  expireYear,
+  securityCode,
+  cardPassword1,
+  cardPassword2,
+}: AddCardFormProps) => {
+  return {
+    cardCompany: cardCompany.value,
+    cardNumber: {
+      first: cardFirstNumber.value,
+      second: cardSecondNumber.value,
+      third: cardThirdNumber.value,
+      fourth: cardFourthNumber.value,
+    },
+    cardOwner: cardOwner.value,
+    expireMonth: expireMonth.value,
+    expireYear: expireYear.value,
+    securityCode: securityCode.value,
+    cardPassword: {
+      first: cardPassword1.value,
+      second: cardPassword2.value,
+    },
+  };
 };
